+++ CHANGELOG.markdown
@@ -4,6 +4,7 @@
 * Expose `HasTypes` class in `Language.Haskell.TH.Lens`
 * Make types of `foldByOf` and `foldMapByOf` more specific to hide implementation details
 * Add Prisms to `Language.Haskell.TH` for new constructors in `template-haskell-2.10`
+* Generalize type of `_FunDep` to an `Iso`
 
 4.12.2
 ------